[IMP] repair_refurbish_repair_stock_move: allow setting a repair order "to refurbish" in the middle of the process.
The stock moves are already created but in the middle the user decides the repair will be refurbish.

Before this commit the stock moves are already created and they are not valid for the new approach. Also
the repair refurbish move is not created.

The solution consist of cancelling existing moves and recreating them

It also covers the case when the repair passes from "to refurbish" to a normal repair